refactor: 당첨 내역 일급 컬렉션으로 리팩토링
